factors out command execution
