+++ commands/WirelessCommand/WiFiManager.js
@@ -15,13 +15,14 @@ function WiFiManager(opts) {
 
 	this.platform = os.platform();
 	this.osConnect = connect[this.platform];
+	// todo - allow the connector to actively check for preconditions, specific OS version support etc
 	this.supported = {
 		getCurrentNetwork: !!(this.osConnect && this.osConnect.getCurrentNetwork),
 		connect: !!(this.osConnect && this.osConnect.connect)
 	};
 
 	this.__cache = undefined;
-};
+}
 
 WiFiManager.prototype.getCurrentNetwork = function(cb) {
 	if (!this.supported.getCurrentNetwork) {
@@ -40,7 +41,6 @@ WiFiManager.prototype.scan = function scan(opts, cb) {
 			return cb(err);
 		}
 		if (dat.length) {
-
 			self.__cache = dat;
 			return cb(null, dat);
 		}
@@ -81,7 +81,7 @@ WiFiManager.prototype.connect = function(opts, cb) {
 		}
 		opts.ssid = ap.ssid;
 		return self.__connect(opts, cb);
-	};
+	}
 };
