+++ commands/WirelessCommand/connect/linux.js
@@ -1,38 +1,14 @@
 var spawn = require('child_process').spawn;
 var wifiCli = '/usr/bin/nmcli';
 
-function runCommand(cmd, args, cb) {
-	var argArray = args.split(' ');
-
-	var s = spawn(cmd, argArray, {
-		stdio: ['ignore', 'pipe', 'pipe']
-	});
-
-	var stdout = '';
-	s.stdout.on('data', function (data) {
-		stdout += data;
-	});
-
-	var stderr = '';
-	s.stderr.on('data', function (data) {
-		stderr += data;
-	});
-
-	s.on('error', function (error) {
-		cb(error, stdout, stderr);
-	});
-
-	s.on('close', function (code) {
-		cb(code, stdout, stderr);
-	});
-}
+var runCommand = require('./executor').runCommand;
 
 function getCurrentNetwork(cb) {
 	var currentNetworkParams = "--terse --fields NAME,TYPE connection show --active";
 
-	runCommand(wifiCli, currentNetworkParams, function (err, stdout, stderr) {
-		if(err || stderr) {
-			return cb(err || stderr);
+	runCommand(wifiCli, currentNetworkParams, function (err, code, stdout, stderr) {
+		if(err || stderr || code) {
+			return cb(err || stderr || code);
 		}
 
 		var wifiType = "802-11-wireless";
@@ -54,8 +30,8 @@ function connect(opts, cb) {
 	function reconnect() {
 		var connectionDoesNotExistError = 10;
 		var reconnectParams = 'connection up id ' + opts.ssid;
-		runCommand(wifiCli, reconnectParams, function (err, stdout, stderr) {
-			if(err == connectionDoesNotExistError) {
+		runCommand(wifiCli, reconnectParams, function (err, code, stdout, stderr) {
+			if(code == connectionDoesNotExistError) {
 				return newConnect();
 			} else if(err || stderr) {
 				return cb(err || stderr);
@@ -71,9 +47,9 @@ function connect(opts, cb) {
 			newConnectParams += ' password ' + opts.password;
 		}
 
-		runCommand(wifiCli, newConnectParams, function (err, stdout, stderr) {
-			if(err || stderr) {
-				return cb(err || stderr);
+		runCommand(wifiCli, newConnectParams, function (err, code, stdout, stderr) {
+			if(err || stderr || code) {
+				return cb(err || stderr || code);
 			}
 
 			cb(null, opts);
